✨ HappyLearning: add class hours tracking and status display
- Bump version to v1.1.0
- Add classHours field to track lesson duration
- Add status label showing learning progress
- Add mutual exclusion for major/elective checkboxes
- Fix typo: UnHAppy → UnHappy
- Use getName() instead of getBrief() for lesson name

+++ OfficalPlugin/SwissKitJ-Plugin-HappyLearning/src/main/java/plugin/swisskit/hpl/dto/LessonDetailVO.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class LessonDetailVO {
     private Long lessonId;
     private String name;
-    private Integer classhour;
+    private Float classhour;
     private Integer letime;
     private Integer isonline;
     private String picfilename;
